Cosmetic changes

+++ abipy/eph/vpq.py
@@ -17,7 +17,7 @@ from scipy.interpolate import interp1d
 #from monty.termcolor import cprint
 from abipy.core.func1d import Function1D
 from abipy.core.structure import Structure
-from abipy.core.kpoints import kpoints_indices, kmesh_from_mpdivs, map_grid2ibz
+from abipy.core.kpoints import kpoints_indices, kmesh_from_mpdivs #, map_grid2ibz
 from abipy.core.mixins import AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter
 from abipy.tools.typing import PathLike
 from abipy.tools.plotting import (add_fig_kwargs, get_axarray_fig_plt, set_axlims, set_visible,
@@ -169,11 +169,11 @@ class VpqFile(AbinitNcFile, Has_Structure, Has_ElectronBands, NotebookWriter):
         e_frohl = r.read_value("e_frohl") # in Ha
 
         d = dict(
-            avg_g = bool(avg_g),
-            e_frohl = e_frohl * abu.Ha_eV,
+            avg_g=bool(avg_g),
+            e_frohl=e_frohl * abu.Ha_eV,
             ngkpt=tuple(ngkpt),
-            inv_k = 1. / np.cbrt(nkbz),
-            invsc_linsize = 1. / np.cbrt(nkbz * self.structure.lattice.volume),
+            inv_k=1. / np.cbrt(nkbz),
+            invsc_linsize=1. / np.cbrt(nkbz * self.structure.lattice.volume),
         )
 
         return d
@@ -245,7 +245,7 @@ class Polaron:
     nq: int            # Number of q-points in B_qnu (including filtering if any).
     bstart: int        # First band starts at bstart.
     bstop: int         # Last band (python convention)
-    erange: double     # Filtering value (in Ha)
+    erange: float     # Filtering value (in Ha)
     varpeq: VpqFile
 
     @classmethod
@@ -549,14 +549,14 @@ class Polaron:
                             ax.set_yscale("log")
                     else:
                         if entry.name == "E_pol":
-                        # Solid line for the *variational* quantity, also put it on top
+                            # Solid line for the *variational* quantity, also put it on top
                             ls, zord = '-', 10
                         else:
-                        # Dashed lines for non-variational, put them below
+                            # Dashed lines for non-variational, put them below
                             ls, zord = '--', 0
 
                         if iax == 1:
-                            energy_like =True
+                            energy_like = True
                             # Plot values linear scale.
                             ax.plot(xs, ys, label=entry.latex, linestyle=ls, zorder=zord)
                         elif iax == 2:
@@ -586,7 +586,6 @@ class Polaron:
                     else:
                         ax.set_title("Log-scale difference")
 
-
         fig.suptitle(self.get_title(with_gaps=True))
         fig.tight_layout()
 
@@ -594,7 +593,7 @@ class Polaron:
 
     @add_fig_kwargs
     def plot_ank_with_ebands(self, ebands_kpath,
-                             ebands_kmesh=None, lpratio: int = 5, with_info = True, with_legend=True,
+                             ebands_kmesh=None, lpratio: int = 5, with_info=True, with_legend=True,
                              with_ibz_a2dos=True, method="gaussian", step="auto", width="auto",
                              nksmall: int = 20, normalize: bool = False, with_title=True, interp_method="linear",
                              ax_mat=None, ylims=None, scale=50, marker_color="gold", marker_edgecolor="gray",
@@ -922,7 +921,6 @@ class Polaron:
 
                     omegas_nu = phbands_qpath._get_non_anal_freqs(nana_dir)
 
-
                 for w, b2 in zip(omegas_nu, b2_interp_state[pstate].eval_kpoint(qpoint), strict=True):
                     w *= units_scale
                     x.append(iq); y.append(w); s.append(_scale * b2)
@@ -1015,7 +1013,6 @@ class Polaron:
                                           label=r"$B^2$(E)", color=marker_color, linewidth=lw_dos,
                                           xfactor=units_scale, yfactor=1/units_scale)
             set_grid_legend(ax, fontsize, xlabel="Arb. unit")
-
 
             # Get mapping BZ --> IBZ needed to obtain the KS eigenvalues e_nk from the IBZ for the DOS
             # Compute B2(E) using only q-points in the IBZ. This is just for testing.
@@ -1280,7 +1277,6 @@ class VpqRobot(Robot, RobotWithEbands):
             print("plot_erange_conv: not enough data fro convergence wrt erange")
 
         return fig_list
-
 
     @add_fig_kwargs
     def plot_kconv(self, nfit: int = 3, spin: int = 0, pstate: int = 0, convby: str = "invsc_linsize",
